[GPFS] Change handle_size to 48 if GPFS returned handle with size 40
Currently for lookup of name (specifically for ..), GPFS may return
a handle with handle_size=40, which is not correct. Expected value
of handle_size for a GPFS handle is 48 bytes.
Till we don't have a fix from GPFS for this issue we will have this
workaround to change the handle_size to 48 in case GPFS returned handle
with handle_size=40.

Change-Id: If547549397d2577301fa752abf6e60f05e11da1c
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>
(cherry picked from commit 894ea80f0a7c189eac5817869595e365e9891b5d)