Fix Ganesha crash by sending stats
"ganesha_stats" with option client_all_ops sends stats for client
specific ops, but only if they are non-zero. In case if "reset"
operation was conducted, and there is no further activity by that
client, then no stats would be sent, leading to crash of Ganesha in
DBUS library code.
To avoid this, work-around is opted in which stats for the first
operation of that protocol will be always sent across. This will
avoid the crash.

Change-Id: Iec81e27d23459374062434b64f2273f97e548e86
Signed-off-by: Sachin Punadikar <psachin@in.ibm.com>
(cherry picked from commit eaa8bc569e3af2d32ed24757790ac15c9cfa2309)